Adds refactored subject page with components folder #1048
